backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
aiofiles==23.2.1
websockets==12.0
pydantic-settings==2.2.1
email-validator==2.1.1
openai==1.14.3
httpx==0.26.0
anyio==3.7.1
assemblyai==0.53.0